+++ cpu/emulator_function.h
@@ -15,11 +15,11 @@ int8_t get_sign_code8(Emulator* emu, int index);
 uint32_t get_code32(Emulator* emu, int index);
 int32_t get_sign_code32(Emulator* emu, int index);
 
-void set_memory8(Emulator* emu, uint32_t address, uint32_t value);
-void set_memory32(Emulator* emu, uint32_t address, uint32_t value);
+void set_memory8(Emulator* emu, uint64_t address, uint64_t value);
+void set_memory32(Emulator* emu, uint64_t address, uint64_t value);
 void set_memory64(Emulator* emu, uint64_t address, uint64_t value);
-uint32_t get_memory8(Emulator* emu, uint32_t address);
-uint32_t get_memory32(Emulator* emu, uint32_t address);
+uint64_t get_memory8(Emulator* emu, uint64_t address);
+uint64_t get_memory32(Emulator* emu, uint64_t address);
 uint64_t get_memory64(Emulator* emu, uint64_t address);
 
 uint8_t get_register8(Emulator* emu, int index);
